+++ shared/CHANGELOG.md
@@ -4,6 +4,24 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.4](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.13.3...@wundergraph/cosmo-shared@0.13.4) (2023-11-20)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
+## [0.13.3](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.13.2...@wundergraph/cosmo-shared@0.13.3) (2023-11-17)
+
+### Bug Fixes
+
+* show latest valid subgraph schema ([#259](https://github.com/wundergraph/cosmo/issues/259)) ([d954b91](https://github.com/wundergraph/cosmo/commit/d954b91bd212ae1a33257c662a4ff8a2ac8c2b56)) (@JivusAyrus)
+
+## [0.13.2](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.13.1...@wundergraph/cosmo-shared@0.13.2) (2023-11-15)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
+## [0.13.1](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.13.0...@wundergraph/cosmo-shared@0.13.1) (2023-11-15)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
 # [0.13.0](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.12.5...@wundergraph/cosmo-shared@0.13.0) (2023-11-10)
 
 ### Features